requirements.txt
python-dotenv>=1.0.1
fastapi~=0.115.2
pydantic~=2.9.2
uvicorn~=0.32.0
pymongo~=4.10.1
motor~=3.6.0
aiohttp~=3.10.10
python-multipart~=0.0.12